requirements.txt
streamlit
pdf2image
Pillow
pdfplumber
pytesseract
requests
